completed the declaration of methods
